prevent glyphIdToCharacterCode from being allocated multiple times.
The method allocated it unconditionally at the beginning using newGlyphIdToCharacterCode
but overwrites it at the end using buildGlyphIdToCharacterCodeLookup. Don't allocate it
at the start of the method but null it. Only explicitly allocate an glyphIdToCharacterCode
object in special case (zero glyphs), otherwise the array gets built at the end of the method.